Tests: implement unit tests (2/2)

+++ src/domains/vehicles-skins/imagery.helper.ts
@@ -3,7 +3,11 @@ import { S3 } from 'aws-sdk';
 
 @Injectable()
 export class ImageryHelper {
-  public async upload(uuid: string, fileBuffer: Buffer): Promise<void> {
+  public async upload(uuid: string, fileBuffer: Buffer): Promise<string | void> {
+    if (process.env.STAGE !== 'production') {
+      return Promise.resolve('Skipping bucket transactions outside production environment');
+    }
+
     const s3: S3 = new S3();
 
     await s3
@@ -15,9 +19,12 @@ export class ImageryHelper {
       .promise();
   }
 
-  public async remove(key: string): Promise<void> {
-    const s3: S3 = new S3();
+  public async remove(key: string): Promise<string | void> {
+    if (process.env.STAGE !== 'production') {
+      return Promise.resolve('Skipping bucket transactions outside production environment');
+    }
 
+    const s3: S3 = new S3();
     await s3.deleteObject({ Bucket: process.env.BUCKET_NAME, Key: key }).promise();
   }
 }
